modules

+++ src/BasicNft.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.18;
 import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 
 contract BasicNft is ERC721 {
-    uint private s_tokenCounter;
+    uint256 private s_tokenCounter;
     mapping(uint256 tokenId => string) private s_tokenIdtoUrl;
 
     constructor() ERC721("Dogie", "DOG") {
@@ -18,9 +18,7 @@ contract BasicNft is ERC721 {
         s_tokenCounter++;
     }
 
-    function tokenURI(
-        uint256 tokenId
-    ) public view override returns (string memory) {
+    function tokenURI(uint256 tokenId) public view override returns (string memory) {
         return s_tokenIdtoUrl[tokenId];
     }
 }
